==> 05-followup-sampling.ipynb <==
# Preparation

import datetime as dt
import re
import json
import time
import pandas as pd
import numpy as np

# An arbitrary integer to use as a sampling seed so the sampling is deterministic and can be re-run
RANDOM_STATE = 788319

frame = pd.read_parquet("secrets/sampling-frame.parquet")
previous_sample = pd.read_table("secrets/sampled_users.tsv").set_index("user_name")
original_strata_targets = pd.read_table('definitions/strata-sample-targets.tsv').set_index('project_group')

The sampling targets (both for actual and follow-up) are dependent on the response rates of the respective wikis. For enwiki, dewiki, and commonswiki, we only sample 75% of the actual target in anticipation of higher response rates (due to a new distribution mechanism). The follow-up sample is only required for enwiki, where we will sampling the remainder of 25% of the actual sampling targets.

strata_targets = original_strata_targets.copy()
strata_targets.iloc[:] = 0.
strata_targets.loc['enwiki', :] = round(original_strata_targets.loc['enwiki', :] * 0.25)
strata_targets

# Remove previously sampled users from frame

frame = frame.drop(previous_sample.index)

# Set sample sizes

# Stratify sample
strata_sizes = frame.groupby(["project_group", "edit_bin"]).size().unstack()

# `strata_targets` has string column names, but `strata_sizes` has categoricals
strata_targets.index = strata_sizes.index
strata_targets.columns = strata_sizes.columns

for i in strata_targets.index:
    for col in strata_targets.columns:
        if strata_targets.loc[i, col] > strata_sizes.loc[i, col]:
            strata_targets.loc[i, col] = strata_sizes.loc[i, col]
            print(f'{i} ({col}) capped at {strata_sizes.loc[i, col]}')

# The targets must not be larger than the strata, or the sampling will hit an error
excessive_targets = (strata_targets > strata_sizes)
assert excessive_targets.any(axis=None) == False

# Set target to 0 for any strata with fewer than 10 users,
# to prevent cases where de-anonymization of data would be too easy
strata_sizes = strata_targets.where(strata_sizes >= 10, other=0)

strata = (
  strata_sizes
  .unstack()
  .reset_index()
  .rename({0: "sample_size"}, axis=1)
  # Ensure columns are ordered in the expected way
  [["project_group", "edit_bin", "sample_size"]]
)

strata.head()

# Pull samples

def sample_stratum(project_group, edit_bin, sample_size):
    """
    returns the sampled users as a data frame
    """
    stratum = frame.query("project_group == @project_group & edit_bin == @edit_bin")
    sample = stratum.sample(n=sample_size, random_state=RANDOM_STATE)
    return sample

f = []

for row in strata.itertuples(index=False):
    f.append(sample_stratum(*row))
    
sampled_users = pd.concat(f)

sampled_users.info()

# Save samples table

sampled_users.to_csv("secrets/followup-sampled-users.tsv", sep="\t")
